backend/pyproject.toml
[project]
name = "yihua-lab-backend"
version = "0.1.0"
description = "Blog backend API powered by Flask + Supabase"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1",
    "flask-cors>=5.0",
    "supabase>=2.11",
    "python-dotenv>=1.0",
    "markdown>=3.7",
    "pygments>=2.18",
    "pyjwt>=2.8",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
]
